refactor: [APP-P2-001] Fix architectural violations
Move IEpubLoader from Domain layer to Application layer to maintain
clean architecture boundaries. Infrastructure concerns like file I/O
should not reside in the domain layer.

Changes:
- Move IEpubLoader interface to Alexandria.Application.Services
- Update all references across codebase
- IBookRepository already correctly positioned in Domain.Repositories

All tests pass: 206 passed, 1 skipped

Closes #2
