+++ src/Alexandria.Application/Features/LoadBook/LoadBookHandler.cs
@@ -1,6 +1,6 @@
+using Alexandria.Application.Services;
 using Alexandria.Domain.Entities;
 using Alexandria.Domain.Errors;
-using Alexandria.Domain.Interfaces;
 using Alexandria.Domain.Services;
 using FluentValidation;
 using MediatR;
