+++ src/Alexandria.Infrastructure/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Alexandria.Application.Features.LoadBook;
+using Alexandria.Application.Services;
 using Alexandria.Domain.Interfaces;
 using Alexandria.Domain.Services;
 using Alexandria.Infrastructure.Caching;
