+++ tests/Alexandria.Application.Tests/Features/LoadBook/LoadBookHandlerTests.cs
@@ -1,4 +1,5 @@
 using Alexandria.Application.Features.LoadBook;
+using Alexandria.Application.Services;
 using Alexandria.Domain.Common;
 using Alexandria.Domain.Entities;
 using Alexandria.Domain.Errors;
